Refactoring des batchs et des requetes ElasticSearch Ajout de l'id des installations dans ElasticSearch
