+++ src/main/java/nosql/workshop/services/InstallationService.java
@@ -71,6 +71,7 @@ public class InstallationService {
                 .and("{$unwind: \"$equipements.activites\"}")
                 .and("{$group: {_id: \"$equipements.activites\", total:{$sum : 1}}}")
                 .and("{$project: {activite: \"$_id\", total : 1}}")
+                .and("{$sort: {total: -1}}")
                 .as(CountByActivity.class).iterator()));
 
         installationsStats.setAverageEquipmentsPerInstallation(
@@ -91,7 +92,7 @@ public class InstallationService {
                 .as(Installation.class));
 
         installationsStats.setTotalCount(this.installations.count());
-        
+
         return installationsStats;
 
     }
